Add participant types and statuses to misbehavior document
Specify the basic types of participants and how they behave under different
incentives. These types can be used in the yellowpaper and design documents to
formalize attacks, misbehaviors and incentive structures.

Additionally, specify the basic statuses of participants in the network. All
behavior can be reduced to the basic behaviors:

- well-behaved: sends the expected, valid messages at the correct times
- inactive: fails to send the expected message
- misbehaving: sends an invalid message