Add staking contract glue to ethereum package
Both TokenStaking and StakingProxy Go bindings were auto-generating, but
there was no glue to instantiate those bindings, pull an address from
the config, and store them on the overall ethereum object. This commit
fills that in, including intermediary functions to access the stake
balance for a given address and to watch for unstake events.