Skip the accusation if the current member is involved
The previous logic assumed that if the current member is involved in the
accusation, we report an error. However, accusation messages contain a
map of all accused members so it's not feasible to filter out certain
members from it. From now on, we just ignore a part of the accusation
message the current member is involved in.