WatchUnstakedFor watches indefinitely
There was a return after the first watched event was received, which
meant the watch call would only work for the first event. The return is
now gone, allowing the watch to go on indefinitely.

There is a broader need for callback-style functions to provide a way to
remove the passed listener when it's no longer needed, but that will be
addressed more broadly across the system in separate work.