Ignore messages that are not appropriate for the current state
Instead of raising an error, we just ignore messages that are not
appropriate for the current state. Error, if returned, means that
something went wrong on our side. We don't want to react with an error
on incorrect behavior of other group members since it's an easy way to
nuke the protocol execution without a security upside.