Glue code between GJKR state machine and network
The code in most parts is based on the one for J-F DKG protocol version.
There is one important difference: we do not switch to the next state
once we receive a message. We always try to swith all members at the
same time once the block counter clock turns.