Encapsulate random-beacon content into one subdir
Since we'll be splitting out a separate repo with the random beacon
yellowpaper, move the joining doc into the random-beacon subdirectory as
index.adoc.

A copy of the glossary is also placed into the random-beacon directory
for now; how to handle this more generally since the glossary is a
project-level glossary will be decided later.