Major update with 4 propositions for stake delegation
This update introduces four different strategies for achieving
stake delegation. It is the result of our previous discussions,
where we have highlighted the impact on the network throughput.
As we need to balance between the operations safety and network
throughput we need to choose the best possible strategy.
Therefore, the following options were proposed:

1. Stake delegation with stake atomicity
2. Stake delegation with multioperation
3. Stake delegation with sub-stake atomicity and limited multi-
operation
4. Stake delegation with somewhat stake atomicity and limited
multioperation

First one, provides lowest network throughput but high operations
safty. Second one, guarantees highest network throughput for the
cost of low operations safety. Next two are balancing between high
throughput and safety.

One remark regarding operations safty is that the term should be
interpreted as a staked assets safety. Highest operations safety
means that it is harder (it requries more time) to harm the owner
by provoking malicious operations and enforcing penalisation of
the owners stake by the stake operator.