Removed t.Parallel() call from block waiter tests
With t.Parallel() in this place, our tests being executed does not
receive correct parameters.

In our case, we have the func(t *testing.T) {...} closure, that is
ran in a go routine invoked inside the t.Run() function. When
calling t.Parallel() the test sends a signal to its parent test to
stop waiting for it, and then the loops continues.
This causes the tc variable to advance to the next tests value.

See "Using goroutines on loop iterator variables" section in Go common
mistakes docs:
https://github.com/golang/go/wiki/CommonMistakes#using-goroutines-on-loop-iterator-variables