Phases that send no messages are now active for 0 blocks
Since we do not wait for any messages and we only execute some
calculations we can speed up the protocol by setting activeBlocks to 0
for those phases. This way, we will proceed to the next phase, straight
after completing all the phase calculations, without waiting for a
block to be mined.