Implement GetOrderedTickets and verifyTicket
GetOrderedTickets fetches the tickets that have been submitted to the
chain that have passed all the checks. We use this in verifyTickets, a
long running loop that gets all the tickets that have passed checks on
chain (every tick, 5 seconds), checks each one for integrity, and
submits a challenge for any ticket that fails a check.