Test ticket ordering, fix implementation
By testing the sort.Stable implementation (that tickets are in ascending
order), a bug with the sort implementation was unvieled: rather than
testing if the tickets were ordered by virtual staker index, it was
comparing the value of the ticket. This led to indeterminate results.
Fixed by sorting by the virtual staker index (as originally intended).