Allow to return error from OnDKGResultPublished function
So far all our watch functions in the chain interface did not return an
error. However, they were also not returning a subscription. If
something was going wrong, the error was logged and callback pased as an
argument was never called. In this case we return a subscription so we
need to return an error as well.