Restructure code
We restructure the code so that we don't constantly trigger a callback,
and race between Submitting Tickets, incomming group results, and
executing DKGs. Rather, submitting tickets is a discrete action that
happens when we see a new relay entry, and joining a group happens when
we see a new group result. We make the following assumption: that the
group selection result contains the event.Entry.