Remove node.StakeID; use node.Staker.ID()
First, we ensure that the Staker interface returns []byte for ID()
rather than a string. Next we update all assumptions in code we had for
the staker being a string (all downstream types). Once we're certain all
of those are replaced, we update the code to use node.Staker.ID() rather
than node.StakerID, which leads to cleaner more maintainable code (and
also a single refernce to ID rather than multiple).