Refactor fetching selected tickets
First, we ensure that the first thing we do is kill the challenge loop.
Next, we add the selected tickets to the group candidate state. During
ticket verification, for every call to GetOrderedTickets that ends up
being successful, we save that to our state. After that process has
ended, and we're checking the on-chain selected tickets, we add a degree
of robustness by referencing the most recent tickets from our state IF a
call to the on-chain method fails.