Capture player index in loop before calling goroutine
The index is actually a single variable updated with each loop
iteration, because all goroutines are bound to that variable, we were
getting the current iteration index at the moment of starting goroutine
and not the iteration index at the moment when we scheduled goroutine to
start. As a result, usually, all players were starting with index equal
to group size.