Use a real protocol identifier for GJKR DKG protocol
We were reading protocol identifier from net.BasicMessage which takes it
from network message where in fact, in network message, we have a
transport identifier, not a protocol identifier.

I have introduced an interface each GJKR message implements and this
interface exposes `SenderID()` function. Once we cast a `net.Message` in
the state machine into appropriate phase message, we can use that
interface to retrieve protocol sender ID.

The `ProtocolSenderID()` function from `net.Message` should be renamed
to `TransportSenderID()` and this will be done in a separate commit.