Updates to beacon on relay entry request and sub
On Relay Entry Request executes group initialization. It is assumed that
beacon will be called by the request function for the first time. The
group will be initialized and ready for the DKG execution.

On Relay Entry Submitted will execute new group generation in pararrel
with a new relay entry generation and submission.