Clean up callback, subscription, and watcher code
Previously, we were using contexts to control unsubscribes. This doesn't
work, as geth ignore's these contexts. Rather, we unsubscribe from the
subscription, close the event channel, and return after error. This
allows us to kill the Watcher and callback if an only if we explicitly
call unsubscribe or receive an error from geth.