Updates to relay entry submitting command
Command to submit relay entry has been updated to generate an entry with
random values.
Additional command `genesis` was added to cover initial entry creation.
It is expected that by running this command initial group will be created.
On clean chain it is required to first run `genesis` command and after
the group is formed `submit` command might be executed.