Fix nasty race in publishing
Deferring the close of the channel resulted in the case where we'd
publish a result and return, while getting a hit on the callback that
publishes to the channel (for the recently submitted dkg result). In
this case, we don't close the channel. It's to be seen if this leaks or
will get garbage collected properly.