Ensure we unsubscribe
Rather than defer the subscription's unsubscribe, we explicitly call it
at all return points, with the execption of the publishing case. In the
case that we are selected to publish, we unsubscribe from the
subscription (avoiding a collision with the subscription for the same
callback nested in SubmitDKGResult) as well as safely closing the
channel before submitting the result (as we'll no longer need to monitor
the callback regardless of the result).