Pass seed to chain and receive it from chain
When submitting relay entry and receiving relay entry submitted event
from chain we should include seed in parameters list. So far, the seed
was ignored and as a result we were receiving `nil` everywhere. Seed is
important for M2, because DKG protocol use seed to calculate H parameter
for Pedersen commitment (G1 hash to point).