Add getSelectedTicketsWithRetry
getSelectedTicketsWithRetry attempts to call
relayChain.GetSelectedTickets at the rate of every tick with exponetial
backoff, within a window of 16seconds. This allows for a bit of
resilinecy within local calls - different nodes competing for the same
resources on the machine will be able to pull this information from geth
at different times, approximately within the window of time.