Begin changing G1 to G2 where applicable
The points G1 and G2 correspond to the BLS signature and the BLS key
pair. Concretely, the private key is a scalar in G2, and the public key
is a point on the curve. Previously, we were using G1 (the signature)
rather than G2 for the public key and public key shares. This is
incorrect. This commit begins moving sharing member (phases 7 and 8 of
gjkr) to the correct curve point.