Added WaitForBlockHeight and BlockHeightWaiter to BlockCounter
Added two new functions to BlockCounter and implemented them for
ethereum and local chain stub. The difference between WaitForBlocks and
BlockWaiter is that new functions allow to set block height precisely
and provide better synchronization guarantees.

I decided to add new functions to allow for gradual migration where it is
needed and possible.