Let signer.GroupPublicKeyBytes use altbn's Compress
Ensure the threshold signer correctly marshals the group public
key to bytes by using the altbn128 method of compressing the point
rather than the direct marshal method on bn256.G2(). Now the signer
membership and the group public key that we see on chain will be in sync
(compressed points that we can compare).