Wait untill the end of Phase 14 duration
Publisher should wait untill Phase 14 time slot comes to an end, even if
he won't be allowed to submit or vote again. It is done so we can
synchronize the phase execution across the members and when it ends we
will call function to get final result from the chain.