Get secret shares indexing right
Every member should be indexed from 1, we should not embed this
information in bls functions - it's the responsibility of the
caller/protocols. The variable `xi` corresponding to the player
index should be set to `i`, where `i` is correctly set to `i+1` (if
necessary) by the protocol / caller.

Tests will fail. We'll fix them separately.