Fix BLS tests, ensuring members are 1-indexed
Remove the embedded knowledge from BLS where we shifted members by 1.
This shouldn't be in functions as it will result in false positives.
Rather, it's a feature that tests will fail if not 1-indexed as it
forces the developer to ensure that they are passing the correct,
shares, keys, and indicies through.