Introduced DkgResultElectedEvent
We need a separate event to signal that DKG result election has been
completed. GroupRegisteredEvent can't be used for this purpose because
not every DKG result selection will result in group registration. In
case we have no honest majority of votes on the leading result, the
entire DKG operation finishes with failure.