First pass at agnostic static key package
In DKG Phase 13 and 14, we'll need to sign the group public key with the
operator's static public key as a proxy for voting. Previously, our
static public key was confined to the network package. This commit
brings that functionality to our top level, allowing it to be used both
in net and in dkg.