Add simple random beacon entry UI
Currently this is an HTML page with inline JS, and an associated Keep
logo SVG file and CSS file. No real design has been applied here, and
not everything is accurate, but this page loaded up on a server that is
also serving up the routes served by the `keep api` command will produce
a way of viewing the latest entry, refreshed every second from the API.

One could also conceive of a server push via WebSocket or similar to
notify of new entries, but for now polling will do.