Ensure compatibility with crypto.VerifySignature
A given signature may be 65 or 64 bytes. If the signature is 65 bytes,
this means it's been signed in such a way to be compatible with
Ethereum. We remove the V from the very end of the array. The correct
size for ECDSA signatures is 64 bytes. We also convert our ecdsa key to
the correct btcec format.