Use crypto.GenerateKey
Rather than manually generating a key from the ecdsa package with a
specific curve, we use go-ethereum/crypto's GenerateKey function which
uses whichever curve go-ethereum determines suitable for the system.
This avoids a whole host of issues around curve incompatability with our
code (and pushes the responsibility to go-ethereum as they do a variety
of things with different curves).