Add two helper types to network public key
NetworkKeyToECDSABytes converts valid network public key's to an ecdsa
public key, as a slice of bytes. This is a necessary conversation method
as we need to go from a network public key, to and ecdsa key for
external consumers. And that key must be as a slice of bytes for
efficient transfer AND processing by external packages (who need not
know about net.key.PublicKey and libp2pcrypto.PubKey).

Also introduce Libp2pKeyToNetworkKey to convert from the libp2pcrypto
interface type to the concrete type necessary in our packages.