Disarmed random beacon loop
We no longer want to have random beacon looping. From now on, we want to
respond on relay entry requests.

1. An external client sends a relay entry request to the chain providing
seed for the request

2. KeepRandomBeacon contract chooses request ID and emits RelayEntryRequested
event

3. Keep clients see RelayEntryRequested and they pick a group from those
already known based on previous entry mod number of groups

4.Keep clients who belong to the group execute threshold signing protocol
over the previous entry value and submits a new relay entry to the chain

5. KeepRandomBeacon contract validates signature over the previous entry
and if it is valid, accepts the new relay entry. When entry gets accepted,
chain emits RelayEntryGenerated and starts new group selection

6. Keep clients react on RelayEntryGenerated by starting the group selection
protocol. When group selection ends, a new group is registered off-chain
and on-chain but nothing else automatically happens.<Paste>