Increased block step for DKG result publisher
We have 3-blocks step for each DKG phase involving message exchange. We
should use similar value for publisher's block step. On local geth,
blocks can be mined relly fast and we sometimes end with a situation
when two publishers (usually P_1 and P_2) tries to submit at the same
moment. This is similar problem as with messages exchange - blocks are
mined too fast locally and we sometimes have not enough time to exchange
all messages.