Keep proto share map keys as uint32
We were previously marshalling gjkr.MemberID and member.Index to bytes
or string. String type was used for map keys because array of bytes
cannot be a map key. We no longer use bytes for serialization and in
this commit we port string map keys to uint32 as well. This way, we have
all sender/receiver proto representations consistent.