Convert signatures map to two slices
Ethereum contract requires signatures and members indices to be provided
in two separate slices. We handle them off-chain as a map. Here we
convert the map to the slices. First slice contains indices of members
who provided signatures. Second slice is a concatenation of signatures.
Members indices and concatenated signatures keep the same sequence.