Introduce operator.Unmarshal
Unmarshal takes raw bytes and produces an uncompressed, operator's
PublicKey. This function assumes the PublicKey's curve is of type
crypto.S256. Our only options are to use this method directly OR to
encode the curve type as a bit in the message (passing it in). We opt
for the simpler method.