DKG group and MessageFiltering interface moved to group package
Previously, Group and MessageFiltering were tightly coupled with GJKR
protocol. We need both group and MessageFiltering after GJKR is done to
publish the result to the chain in DKG phase 13 and 14.

Extracting Group and MessageFiltering to a separate package will allow
us to reuse this logic in DKG phase 13 and 14 state machine to filter
out inactive group members.