Validate key used to sign DKG result against the one identifying client
Network layer determines message sender's public key based on network
client's pinned identity. Sender can not use other public key than the
one it is identified with in the network and sender must posses the
private key - each network message is signed with it.

Network layer rejects message with incorrect signature or altered
public key. At this point we know that the sender public key as presented
in the network net.Message is the correct one.

We need to compare that key with one used to produce a signature over
DKG result hash. If those keys don't match, it means that an incorrect
key was used to sign DKG result hash and the message should be rejected.