WatchDKGResultPublishedEvent in OnGroupRegistered
Previously we were watching for group public key submitted. Since the
off-chain part does no longer submit the key to the chain, and instead
it submits DKG result we need to watch for DKG result submitted. When
DKG result is accepted by chain (and event emitted) it means that the
result is valid and accepted as the final one.