Removed gjkr.Result.Equals function
Result contains not only Group and GroupPublicKey reference but also
GroupPrivateKeyShare that was ignored in Equals. Hence, Equals was
comparing only public parameters and was potentially misleading.

The other reason for deleting Equals function is that it was not used in
any other place but tests where we can compare Group and GroupPublicKey
manually.

In the new implementation of DKG P13/14 we do not have to compare
results.