Represent block number/height as uint64 in block waiter
So far we were representing block height in block waiter as int. This is
not perfect since int has platform dependent length and allows for
negative values. The other important aspect is that go-ethereum
represents block height as uint64. Althought we don't our chain
interface to be tightly coupled with Ethereum, choosing uint64 over just
int is safer regardless of what chain do we have under the hood.