Include group public key in RelayEntryRequested event
It is possible that because of a delay in block propagation two Keep
clients can have a different view of the chain and see a different
number of groups registered in the keep group contract.

Group which is going to produce a new relay entry is chosen by its
index, based on previous relay entry modulo number of groups.

As mentioned, two Keep clients can have a different view of
the chain and the modulus used by them can be different. This can lead to
chosing two different groups by Keep clients for the same relay request.

To fix this problem we specify which group should answer the request by
specifying group public key in the event.