refactor configuration file and environment parsing using micro/go-config
we need a better way to manage configurations for the Keep client
that can parse both parameters set in a configuration file and
give us the option to override these values using environment
variables at runtime. The main reason for the need for env variables
is the deployment of Keep clients via Kubernetes.