Block number precise state machine
From now on, state machine starts execution exactly at startBlock passed
as a parameter to Execute. It also takes the previous state's ending
block as an input parameter to active time of the new state. This way,
we don't loose any new blocks along the way and state machine start and
end time can be precisely evaluated based on the number of states and
their active time.