Fetch selected tickets/participants right after challenge timeout passed
When ticket submission started at block 10, and challenge timeout is 4,
we should be able to fetch tickets when block 14 has been sealed -
candidates could challenge in block 11, 12, 13, and 14. Actual behaviour
was different than expected - Keep group contract allowed to fetch
selected tickets no earlier than after block 15 has been sealed.

block.number in view function is the latest block, so the one that was
sealed, not the one currently being mined.

block.number for function executed inside transaction is evaluated at
the moment when this transaction is mined, so it's block including the
transaction.

Say we started at _submissionStart = 10 and _timeoutChallenge = 4.
When block 14 has been sealed, we can no longer submit tickets
(because transaction will be processed in block 15 or later) and we
should be able to get those selected tickets. However, for a view
function, block.number is the latest sealed block, so it's 14.

That's why we should use `>=` instead of `>`.