Abstracted out State.DelayBlocks function
We delay state initialization to give all other cooperating state
machines a chance to perform state transition. For example, when we
exchange messages with other clients we need to be sure we are able to
receive messages for the given state (our state machine is at that
state) before we call Initialize.

I have abstracted out DelayBlocks into a function on State. This will
give us a chance to set 0 for those states that does not require any
delay. This should speed up state machine execution.