Treat DKG result publication phase as "silent" one
We do not exchange any messages in the result publication phase. We
publish result to the chain but it's an action blocking all group
members for the same time - members exit when the first valid result is
accepted by the chain. How long it takes depends on the block step and
group size.