Check if result for request ID was published, and not who published it
We do not have voting in the new version of phase 14 and we do not care
which group member published the result. All we need to assert is
whether the result for the given request ID has been already published.

If result is valid and accepted by chain, then it's considered as final.