First pass at an signing integration test
Now that we're leaving signing to go-libp2p-pubusb, we still want to
test that messages can't be tampered with. This proves difficult as we
don't control the bits in libp2p-pubsub that would allow us to test if a
message was tampered with AND we don't expose the bits in our own code
to manage direct connections.