Only one DKT result with the given request ID is allowed
It is no longer allowed to publish more than one DKG result with the
given request ID. In the previous phase 14 approach we were publishing
multiple results and were voting on them. This commit updates the
behaviour of local chain stub so that it's aligned with the new phase 14
solidity implementation.