Add contract generation code
The contract generator is built using golang.org/text/template, like the
async/promise generator. It assembles information both from the parsed
Go ABI structures and some supplemental information from the raw ABI
JSON to produce functions that access const and non-const contract
methods, as well as providing EventSubscription-based watcher methods
for events.

As much string generation as possible is done inside the Go code so as
to minimize the complexity of the templates.

The contract generator lives in contract.go, and gen.go is gone for now.