isGroupRegistered changed to isStaleGroup
Stale group is an expired group which is no longer performing any
operations. It is important to understand that an expired group may
still perform some operations for which it was selected when it was
still active. We consider a group to be stale when it's expired and when
its expiration time and potentially executed operation timeout are both
in the past.

Stale group can be safely removed from the memory of an off-chain client
(this will be implemented in a separate commit).