Add contract/%.go to Makefile, with abigen files required
This builds contract bindings in contract/ based on the raw ABI bindings
that are built by abigen. These bindings require the abigen bindings to
be built first, and rely on the generated ABI JSON files as well.

Contract files are only generated for *ImplV1 contracts currently, and
all of these contracts are captured in ${contract_files}, which is used
to drive the `gen_contract_go` rule, and through that the `all` rule.