Make event success func handlers get distinct names
Because these exist in the package namespace, they can occasionally
clash (e.g. if two contracts are based on the same core lib and support
similar functionality like ownership transfers). To disambiguate, the
success function's type alias is prefixed with the contract name, in
addition to including the event's name.