Non-expired group cannot be marked as stale
Group needs to be marked as expired (_expiredOffset must be updated) so
that we can even consider marking it as stale. It is important to bear
in mind that not all groups that qualify as expired must be marked as
such - `selectGroup` may decide to leave group as active if minimum
active group threshold is not satisfied.