Move ErrorResolver and DecryptKeyFile to ethutil package
This package is separate from pkg/chain/ethereum, and is referenced from
generated contracts, which will need to be referenced from
pkg/chain/ethereum to fill out the chain interfaces.

It also more clearly delineates the ethereum package as where the
interface is defined, while the ethutil package offers utilities for
interacting with ethereum.