Swap out hardcoded duration for caller defined
The caller should be able to configure a reasonable duration given the
usecase. In our tests, we'll be leveraging our local network and need
only a second in between checks. In production (using libp2p), 1 second
would be too aggressive. Rather, we'll want this checks happening every
minute or so.