Include sender's public key in the message send via local channel
Our local stub channel implementation did not include sender's public
key in the message. As a result, the message was dropped by DKG protocol
phase 14 when testing with a local channel.

The implementation proposed in this commit allows to specify network and
operator key when connection to stub network and stub chain. Since key
is associated with provider, all threads using the same provider will
use the same key. As a result, the current implementation behaves just
like multiple virtual stakers runinng on the same node.

We may want to enhance this solution in future but it's good enough for
now.