Separate peer key check from stake check
A failure in the stake check (network / rpc issue) should not result in
a disconnect as this can cause cascading disconnects in two separate
systems. Rather, we only want to disconnect when the peer's key material
is incorrect or if the minimum stake is not met. We separate out these
processes to ensure we're only disconnecting from peers that have
incorrect key material, or when they fail the min stake check.

Also, ensure we defer the stake monitor lock rather than explicitly
calling unlock multiple times.