Factor contract parsing out of gen/contract.go
contract.go had become a bit of a mega-tangle, so now code that takes
care of looking at ABI and payable info and generating the data
structures that are emitted into templates is all in
contract_parsing.go. contract.go only deals with loading and parsing
templates, as well as executing templates and writing their results.

Also included: some clearer or more detailed function docs in a few
places.