Added composableArgChecker two such checkers
composableArgChecker allows for cli.BeforeFuncs to be composed for the
purposes of checking different interactions between flags. It provides
an andThen function that allows chaining these checkers.

Also included here are two checkers, one to check for the value flag
being set for commands that will require it, and one to check that block
and transaction are not specified when the submit flag is specified.

Tests come along for the ride!