Updated the terminology from gateway to vendor
Gateway term was quite problematic because it wasn't clean if the
responsibility of gateway is just to point to the correct factory or,
maybe, the responsibility of gateway is to create new keeps (delegating
the work to the factory).

We do not want developers to write a lot of boilerplate code so we
decided that gateways should be used to open new keeps but to do that,
their naming should be improved. And then we came up with the name:
VENDOR.