Respect CommandCallable when generating const commands
CommandCallable reflects whether or not a given method can be called
from the command-line, currently based on whether there is a known way
to parse its parameters from command-line parameters. The generator
correctly does not generate the Go method to call the contract method,
but it still generates the cli.Command entry. This would fail
compilation, but since all of our contract methods are currently command
callable it didn't manifest.

CommandCallable now filters the generation of the cli.Command entries as
well.