UnregisterStaleGroups does no longer return an error
We start this function as `go groupRegistry.UnregisterStaleGroups()` in
beacon.go and we ignore errors there. There is really nothing else we
can do than logging an error in case something wrong happened in
`UnregisterStaleGroups`. To avoid losing the errors, we now log them in
`UnregisterStaleGroups` and this function no longer return any error.