Publish genesis relay entry with 0 signing id
If the genesis relay entry could be requested from service contract,
it would be published with requestId = 1 and signingId = 1. Since it
can't be requested by service contract we should at least make sure that
in the basic deployment setup of one service contract and one operator
contract counters in both contracts match with each other - this will
help debugging potential issues. Having the ID set to 0 is
also a nice indicator that it's a seed entry.