isGroupSelectionInProgress replaced with isGroupRegistered
There is a very small possibility that one group selection will happen
immediatelly after another. In case of a delay in transition to the
submission state of one or more of the signers, they may not receive a
notification about published result from the previous group selection
they took a part in. If such situation happens and another group
selection gets triggered immediatelly, they will wait for the result to
be published because isGroupSelectionInProgress will return true.

Much better solution is to check if group with the given public key is
registered on-chain instead of checking if <any> group selection is
currently in progress.