Updated client code to do not use signingId for DKG result submission
In previous commits, on-chain contract was updated to do not accept
signing ID during DKG result submission. Here, off-chain client is
updated to do not pass signing ID when submitting DKG result.

The reason for this change is that it can be only one DKG executing at
the moment, so any execution ID is not really needed.