Add basic support for env-based log level adjustment
The LOG_LEVEL environment variable can be used to set the log level of
the whole executable, or of particular subsystems. For the whole
executable, LOG_LEVEL=debug will do the trick. For particular
subsystems, a space-delimited set of =-delimited pairs specify log
levels for these subsystems. For example:

    LOG_LEVEL="dht=debug bootstrap=info swarm2=fatal"
    LOG_LEVEL="dht=info"

Supported log levels are the ones supported by whyrusleeping/go-logging:

    critical
    error
    warning
    notice
    info
    debug