Support subsystem prefix wildcards in LOG_LEVEL
Log level directive parsing is now split into two functions, and
log-level-only is not special-cased. Instead, each space-separated
directive is now checked as being a global log level or a
subsystem-specific log level. Subsystem-specific log levels are
additional checked for a * suffix. If a subsystem name ends in *, all
subsystems that start with the remaining name are set to the specified
log level.

Directives are parsed in order, so all subsystems can be set to debug
and then a handful cranked down with a directive like:

    LOG_LEVEL="debug keep*=info swarm2=error"